+++ server/app/controllers/applicant/FileUploadController.java
@@ -1,4 +1,4 @@
-package controllers.fileupload;
+package controllers.applicant;
 
 import static com.google.common.base.Preconditions.checkNotNull;
 import static java.util.concurrent.CompletableFuture.completedFuture;
