+++ server/test/controllers/applicant/FileUploadControllerTest.java
@@ -1,4 +1,4 @@
-package controllers.fileupload;
+package controllers.applicant;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.Mockito.mock;
